+++ webapp/tests/unit/user/settings-plugin-fields.test.js
@@ -5,8 +5,8 @@
  *                  syncSettingsPluginFieldFromElement() tagInput branch, and
  *                  renderPluginCards() initAll() call after DOM insertion.
  * @file            webapp/tests/unit/user/settings-plugin-fields.test.js
- * @version         1.6.41
- * @release         2026-04-20
+ * @version         1.6.42
+ * @release         2026-04-21
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
@@ -33,7 +33,9 @@ global.HTMLElement = dom.window.HTMLElement;
 global.Event = dom.window.Event;
 
 const jpulseCommonPath = path.join(process.cwd(), 'webapp/view/jpulse-common.js');
-const jpulseCommonContent = fs.readFileSync(jpulseCommonPath, 'utf8');
+let jpulseCommonContent = fs.readFileSync(jpulseCommonPath, 'utf8');
+// W-185: strip the unquoted subtree-embed token (would otherwise break JS parse when loaded raw)
+jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.controller\.handlebar\.date\.fromNow\}\}/g, '{}');
 
 const context = vm.createContext(window);
 vm.runInContext(jpulseCommonContent, context);