+++ webapp/tests/unit/utils/jpulse-ui-heading-anchors.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit Tests for jPulse.UI.headingAnchors (W-118)
  * @description     Tests for heading anchor links feature: slugify, ID generation, link creation
  * @file            webapp/tests/unit/utils/jpulse-ui-heading-anchors.test.js
- * @version         1.6.41
- * @release         2026-04-20
+ * @version         1.6.42
+ * @release         2026-04-21
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
@@ -69,6 +69,8 @@ jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.view\.ui\.headingAn
 jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.view\.ui\.headingAnchor\.linkFailed\}\}/g, 'Failed to copy link');
 jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.view\.ui\.headingAnchor\.linkToSection\}\}/g, 'Link to %SECTION%');
 jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.view\.ui\.headingAnchor\.copyLinkTitle\}\}/g, 'Copy link to clipboard');
+// W-185: strip the unquoted subtree-embed token (would otherwise break JS parse when loaded raw)
+jpulseCommonContent = jpulseCommonContent.replace(/\{\{i18n\.controller\.handlebar\.date\.fromNow\}\}/g, '{}');
 
 // Execute the code in the window context
 const vm = require('vm');